all test fixed
